Move service out of ProviderInterface to resolve name spacing
