+++ app/lib/allowed_cross_namespace_usage.rb
@@ -8,5 +8,4 @@ module AllowedCrossNamespaceUsage
   VendorAPIApplicationPresenter = VendorAPI::ApplicationPresenter
   RegisterAPISingleApplicationPresenter = RegisterAPI::SingleApplicationPresenter
   VENDOR_API_VERSION = VendorAPI::VERSION
-  ProviderInterfaceDeclineOrWithdrawApplication = ProviderInterface::DeclineOrWithdrawApplication
 end
